Modularise the command module for each chain
