+++ node/cli/src/command/mod.rs
@@ -1,0 +1,8 @@
+#[cfg(feature = "solo")]
+pub mod solo;
+
+#[cfg(feature = "bajun")]
+pub mod bajun;
+
+#[cfg(feature = "ajuna")]
+pub mod ajuna;
