+++ node/cli/src/lib.rs
@@ -16,6 +16,7 @@
 
 mod cli;
 mod command;
+
 mod command_helper;
 
 pub use cli::*;
